chore: linting

+++ contracts/CitadelMinter.sol
@@ -14,6 +14,7 @@ import "./lib/GlobalAccessControlManaged.sol";
 
 import "../interfaces/citadel/ISupplySchedule.sol";
 import "../interfaces/citadel/IxCitadel.sol";
+
 /**
 Supply schedules are defined in terms of Epochs
 
@@ -27,18 +28,25 @@ Epoch {
 contract CitadelMinter is GlobalAccessControlManaged {
     using SafeMathUpgradeable for uint256;
 
-    bytes32 public constant CONTRACT_GOVERNANCE_ROLE = keccak256("CONTRACT_GOVERNANCE_ROLE");
-    bytes32 public constant POLICY_OPERATIONS_ROLE = keccak256("POLICY_OPERATIONS_ROLE");
+    bytes32 public constant CONTRACT_GOVERNANCE_ROLE =
+        keccak256("CONTRACT_GOVERNANCE_ROLE");
+    bytes32 public constant POLICY_OPERATIONS_ROLE =
+        keccak256("POLICY_OPERATIONS_ROLE");
 
     ISupplySchedule public supplySchedule;
     IERC20Upgradeable public citadelToken;
     IxCitadel public xCitadel;
 
-    uint constant MAX_BPS = 10000;
+    uint256 constant MAX_BPS = 10000;
 
     event SupplyScheduleSet(ISupplySchedule supplySchedule);
 
-    function initialize(address _gac, ISupplySchedule _supplySchedule, IERC20Upgradeable _citadelToken, IxCitadel _xCitadel) external initializer {
+    function initialize(
+        address _gac,
+        ISupplySchedule _supplySchedule,
+        IERC20Upgradeable _citadelToken,
+        IxCitadel _xCitadel
+    ) external initializer {
         __GlobalAccessControlManaged_init(_gac);
         supplySchedule = _supplySchedule;
         citadelToken = _citadelToken;
@@ -47,26 +55,36 @@ contract CitadelMinter is GlobalAccessControlManaged {
         emit SupplyScheduleSet(_supplySchedule);
     }
 
-    function mintAndDistribute(uint marketcap, uint treasury, uint supplyStakedBps, uint supplyLockedBps) external onlyRole(POLICY_OPERATIONS_ROLE) gacPausable {
+    function mintAndDistribute(
+        uint256 marketcap,
+        uint256 treasury,
+        uint256 supplyStakedBps,
+        uint256 supplyLockedBps
+    ) external onlyRole(POLICY_OPERATIONS_ROLE) gacPausable {
         // The minted tokens allocated to funding will be sent to the policy operations manager
-        address policyDestination = gac.getRoleMember(POLICY_OPERATIONS_ROLE, 0);
+        address policyDestination = gac.getRoleMember(
+            POLICY_OPERATIONS_ROLE,
+            0
+        );
 
-        uint toMint = supplySchedule.getMintable();
+        uint256 toMint = supplySchedule.getMintable();
         // citadelToken.mint(toMint);
 
-        uint toStakersBps = MAX_BPS;
-        uint toLockersBps = MAX_BPS;
-        uint toFundingBps = MAX_BPS;
+        uint256 toStakersBps = MAX_BPS;
+        uint256 toLockersBps = MAX_BPS;
+        uint256 toFundingBps = MAX_BPS;
 
-        uint toFundingAmount = toMint.mul(toFundingBps).div(MAX_BPS);
-        uint toLockersAmount = toMint.mul(toLockersBps).div(MAX_BPS);
+        uint256 toFundingAmount = toMint.mul(toFundingBps).div(MAX_BPS);
+        uint256 toLockersAmount = toMint.mul(toLockersBps).div(MAX_BPS);
 
         // Round down to stakers
-        uint toStakersAmount = toMint.sub(toFundingAmount).sub(toLockersAmount);
+        uint256 toStakersAmount = toMint.sub(toFundingAmount).sub(
+            toLockersAmount
+        );
 
         // Send funder amount to policy operations for distribution
         // citadelToken.transfer(toFundingAmount, policyDestination);
-    
+
         // Auto-compound staker amount into xCTDL
         // citadelToken.transfer(toStakersAmount, xCitadel);
 
@@ -75,9 +93,12 @@ contract CitadelMinter is GlobalAccessControlManaged {
     }
 
     /// @dev Highest governance level may swap out the supply schedule contract.
-    function setSupplySchedule(ISupplySchedule _newSupplySchedule) external onlyRole(CONTRACT_GOVERNANCE_ROLE) gacPausable {
+    function setSupplySchedule(ISupplySchedule _newSupplySchedule)
+        external
+        onlyRole(CONTRACT_GOVERNANCE_ROLE)
+        gacPausable
+    {
         supplySchedule = _newSupplySchedule;
         emit SupplyScheduleSet(_newSupplySchedule);
     }
-
 }
